# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Сервер мессенджера на Go. Поддерживает: личные и групповые чаты, голосовые и видеозвонки, stories, голосовые комнаты, каналы, стикеры, опросы, ботов, геолокацию, кастомные эмодзи, эффекты сообщений, избранное и многое другое.",
        "title": "Chat Messenger API",
        "termsOfService": "http://localhost:8080/terms",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080/support",
            "email": "support@chatserver.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/channels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chatdomain.ChatResponse"
                            }
                        }
                    }
                }
            }
        },
        "/channels/subscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "parameters": [
                    {
                        "description": "Channel ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "channelId": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/unsubscribe": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "parameters": [
                    {
                        "description": "Channel ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "channelId": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/{id}/subscribed": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    }
                }
            }
        },
        "/channels/{id}/subscribers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/channeldomain.ChannelSubscriber"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/{id}/subscribers/{userId}/role": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target user ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role: admin or subscriber",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "role": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "channeldomain.ChannelSubscriber": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "subscribedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "chatdomain.ChatResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastMessage": {
                    "$ref": "#/definitions/messagedomain.MessageResponse"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userdomain.UserResponse"
                    }
                },
                "type": {
                    "$ref": "#/definitions/chatdomain.ChatType"
                },
                "unreadCount": {
                    "type": "integer"
                }
            }
        },
        "chatdomain.ChatType": {
            "type": "string",
            "enum": [
                "private",
                "group",
                "channel"
            ],
            "x-enum-varnames": [
                "ChatPrivate",
                "ChatGroup",
                "ChatChannel"
            ]
        },
...
        "messagedomain.MessageResponse": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "chatId": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "duration": {
                    "type": "integer"
                },
                "edited": {
                    "type": "boolean"
                },
                "effect": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "fileUrl": {
                    "type": "string"
                },
                "forwardFrom": {
                    "$ref": "#/definitions/userdomain.UserResponse"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "locationTitle": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "mimeType": {
                    "type": "string"
                },
                "pinned": {
                    "type": "boolean"
                },
                "reactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/messagedomain.Reaction"
                    }
                },
                "readBy": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userdomain.UserResponse"
                    }
                },
                "replyTo": {
                    "$ref": "#/definitions/messagedomain.MessageResponse"
                },
                "sender": {
                    "$ref": "#/definitions/userdomain.UserResponse"
                },
                "type": {
                    "$ref": "#/definitions/messagedomain.MessageType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "messagedomain.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "file",
                "gif",
                "voice",
                "video",
                "video_circle",
                "audio",
                "system",
                "location"
            ],
            "x-enum-varnames": [
                "MessageText",
                "MessageImage",
                "MessageFile",
                "MessageGif",
                "MessageVoice",
                "MessageVideo",
                "MessageVideoCircle",
                "MessageAudio",
                "MessageSystem",
                "MessageLocation"
            ]
        },
...
        "messagedomain.Reaction": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "messageId": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/userdomain.UserResponse"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ERROR_CODE"
                },
                "error": {
                    "type": "string",
                    "example": "error description"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "userdomain.UserResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "lastSeen": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите JWT токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}